[Feat] #18 - View와 VC 분리
