--- ExchangeRate/ExchangeRate/Presentation/ExchangeRate/View/View.swift
@@ -1,8 +1,0 @@
-//
-//  View.swift
-//  ExchangeRate
-//
-//  Created by 천성우 on 4/16/25.
-//
-
-import Foundation
