+++ ExchangeRate/ExchangeRate/Presentation/ExchangeRate/ViewController/ExchangeRateViewController.swift
@@ -19,20 +19,18 @@ class ExchangeRateViewController: BaseViewController {
 
     // MARK: - UI Components
 
-    private let searchBar = UISearchBar()
-    private let tableView = UITableView()
-    private let emptyLabel = UILabel()
+    private let contentView = ExchangeRateView()
     
     // MARK: - View Life Cycle
 
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
+    override func loadView() {
+          view = contentView
+      }
         
     override func bindViewModel() {
         viewModel.action?(.fetch)
 
-        searchBar.rx.text.orEmpty
+        contentView.searchBar.rx.text.orEmpty
             .skip(1)
             .bind(onNext: { [weak viewModel] query in
                 viewModel?.action?(.search(query))
@@ -40,7 +38,7 @@ class ExchangeRateViewController: BaseViewController {
             .disposed(by: disposeBag)
 
         viewModel.state.filteredItems
-            .bind(to: tableView.rx.items(
+            .bind(to: contentView.tableView.rx.items(
                 cellIdentifier: "ExchangeRateTableViewCell",
                 cellType: ExchangeRateTableViewCell.self
             )) { _, item, cell in
@@ -51,59 +49,14 @@ class ExchangeRateViewController: BaseViewController {
         viewModel.state.filteredItems
             .map { !$0.isEmpty }
             .observe(on: MainScheduler.instance)
-            .bind(to: emptyLabel.rx.isHidden)
+            .bind(to: contentView.emptyLabel.rx.isHidden)
             .disposed(by: disposeBag)
-        
-        tableView.rx.modelSelected(ExchangeRateItemDisplay.self)
+
+        contentView.tableView.rx.modelSelected(ExchangeRateItemDisplay.self)
             .subscribe(onNext: { item in
-                print("선택한 셀: 통화코드 - \(item.code), 국가 - \(item.country), 환율 - \(item.rate)")
+                print("선택한 셀: \(item.code) - \(item.country)")
             })
             .disposed(by: disposeBag)
-    }
-
-    // MARK: - Set UIComponents
-    
-    override func setStyles() {
-        view.backgroundColor = .white
-        
-        searchBar.do {
-            $0.placeholder = "통화 검색"
-        }
-        
-        tableView.do {
-            $0.register(ExchangeRateTableViewCell.self, forCellReuseIdentifier: "ExchangeRateTableViewCell")
-            $0.separatorStyle = .none
-            $0.rowHeight = 60
-        }
-        
-        emptyLabel.do {
-            $0.text = "검색 결과 없음"
-            $0.textAlignment = .center
-            $0.textColor = .gray
-            $0.font = .systemFont(ofSize: 16, weight: .medium)
-            $0.isHidden = true
-        }
-    }
-    
-    // MARK: - Layout Helper
-
-    
-    override func setLayout() {
-        view.addSubviews(searchBar, tableView, emptyLabel)
-
-        searchBar.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide)
-            $0.leading.trailing.equalToSuperview()
-        }
-
-        tableView.snp.makeConstraints {
-            $0.top.equalTo(searchBar.snp.bottom)
-            $0.leading.trailing.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-
-        emptyLabel.snp.makeConstraints {
-            $0.center.equalToSuperview()
-        }
     }
 
 }
